+++ toad/nn/trainer/__init__.py
@@ -1,3 +1,4 @@
 from .history import History
-from .earlystopping import EarlyStopping
+from .callback import callback
+from .earlystop import earlystopping
 from .trainer import Trainer