+++ toad/utils/__init__.py
@@ -1,2 +1,3 @@
 from .func import *
 from .decorator import *
+from .progress import Progress